+++ CHANGELOG.md
@@ -135,6 +135,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   constant that resolves to the primary shortcut modifier on the target
   platform.
 - New feature flag `serde` enables serialization support for some types.
+- `WindowHandle::execute` executes a function on the window's thread providing
+  access to an `EventContext`. This can be used to gain access to the window
+  directly, including getting a reference to the underlying winit Window.
 
 
 [139]: https://github.com/khonsulabs/cushy/issues/139